Prog II

+++ Prog_II/src/Ex5_Heranca_Veiculo/Aviao.java
@@ -5,7 +5,7 @@ public class Aviao extends VeiculoAereo {
 
     public Aviao(String fabricante, String modelo, String pais, double posicaoX,
                  double posicaoY, double valor, int passageiro, String codigo,
-                 int posicaoZ, String empresa, int acentos) {
+                 double posicaoZ, String empresa, int acentos) {
         super(fabricante, modelo, pais, posicaoX, posicaoY, valor, passageiro, codigo, posicaoZ, empresa);
         this.velocidadeMaxima = acentos;
     }
