+++ Prog_II/src/Ex5_Heranca_Veiculo/Submarino.java
@@ -1,22 +1,22 @@
 package Ex5_Heranca_Veiculo;
 
 public class Submarino extends VeiculoMaritimo{
-    private int posixaoZ;
+    private double posixaoZ;
     private double profundidadeMaxima;
 
     public Submarino(String fabricante, String modelo, String pais, double posicaoX,
                      double posicaoY, double valor, int passageiro, String codigo,
-                     String empresa, int posixaoZ, double profundidadeMaxima) {
+                     String empresa, double posixaoZ, double profundidadeMaxima) {
         super(fabricante, modelo, pais, posicaoX, posicaoY, valor, passageiro, codigo, empresa);
         this.posixaoZ = posixaoZ;
         this.profundidadeMaxima = profundidadeMaxima;
     }
 
-    public int getPosixaoZ() {
+    public double getPosixaoZ() {
         return posixaoZ;
     }
 
-    public void setPosixaoZ(int posixaoZ) {
+    public void setPosixaoZ(double posixaoZ) {
         this.posixaoZ = posixaoZ;
     }
